+++ rollup.npm.js
@@ -53,7 +53,7 @@ export default [
     globals,
     output: { file: 'dist/react.js', format: 'umd' },
     name: 'Client',
-    plugins: plugins,
+    plugins,
   },
 
   {
@@ -62,7 +62,7 @@ export default [
     globals,
     output: { file: 'dist/client.js', format: 'umd' },
     name: 'Client',
-    plugins: plugins,
+    plugins,
   },
 
   {
@@ -71,14 +71,23 @@ export default [
     globals,
     output: { file: 'dist/react-native.js', format: 'umd' },
     name: 'ReactNativeClient',
-    plugins: plugins,
+    plugins,
   },
 
   {
     input: 'packages/core.js',
     output: { file: 'dist/core.js', format: 'umd' },
     name: 'Core',
-    plugins: plugins,
+    plugins,
+  },
+
+  {
+    input: 'packages/ai.js',
+    external: Object.keys(globals),
+    output: { file: 'dist/ai.js', format: 'umd' },
+    name: 'AI',
+    globals,
+    plugins,
   },
 
   {
@@ -87,7 +96,7 @@ export default [
     globals,
     output: { file: 'dist/ui.js', format: 'umd' },
     name: 'UI',
-    plugins: plugins,
+    plugins,
   },
 
   // UMD and ES versions.