+++ src/client/client.js
@@ -9,7 +9,7 @@
 import { createStore } from 'redux';
 import * as ActionCreators from '../core/action-creators';
 import { Multiplayer } from './multiplayer/multiplayer';
-import { createGameReducer } from '../core/reducer';
+import { CreateGameReducer } from '../core/reducer';
 
 /**
  * createEventDispatchers
@@ -19,12 +19,19 @@ import { createGameReducer } from '../core/reducer';
  * @param {object} store - The Redux store to create dispatchers for.
  * @param {string} playerID - The ID of the player dispatching these events.
  */
-export function createEventDispatchers(eventNames, store, playerID) {
+export function createEventDispatchers(
+  eventNames,
+  store,
+  playerID,
+  credentials
+) {
   let dispatchers = {};
   for (let i = 0; i < eventNames.length; i++) {
     const name = eventNames[i];
     dispatchers[name] = function(...args) {
-      store.dispatch(ActionCreators.gameEvent(name, args, playerID));
+      store.dispatch(
+        ActionCreators.gameEvent(name, args, playerID, credentials)
+      );
     };
   }
   return dispatchers;
@@ -37,12 +44,14 @@ export function createEventDispatchers(eventNames, store, playerID) {
  * @param {Array} moveNames - A list of move names.
  * @param {object} store - The Redux store to create dispatchers for.
  */
-export function createMoveDispatchers(moveNames, store, playerID) {
+export function createMoveDispatchers(moveNames, store, playerID, credentials) {
   let dispatchers = {};
   for (let i = 0; i < moveNames.length; i++) {
     const name = moveNames[i];
     dispatchers[name] = function(...args) {
-      store.dispatch(ActionCreators.makeMove(name, args, playerID));
+      store.dispatch(
+        ActionCreators.makeMove(name, args, playerID, credentials)
+      );
     };
   }
   return dispatchers;
@@ -54,16 +63,19 @@ export function createMoveDispatchers(moveNames, store, playerID) {
 class _ClientImpl {
   constructor({
     game,
+    ai,
     numPlayers,
     multiplayer,
     socketOpts,
     gameID,
     playerID,
+    credentials,
     enhancer,
   }) {
     this.game = game;
     this.playerID = playerID;
     this.gameID = gameID;
+    this.credentials = credentials;
 
     let server = undefined;
     if (multiplayer instanceof Object && 'server' in multiplayer) {
@@ -73,11 +85,28 @@ class _ClientImpl {
 
     this.multiplayer = multiplayer;
 
-    const GameReducer = createGameReducer({
+    this.reducer = CreateGameReducer({
       game,
       numPlayers,
       multiplayer,
     });
+
+    if (ai !== undefined && multiplayer === undefined) {
+      this.bot = new ai.bot({ game, ...ai });
+
+      this.step = () => {
+        const state = this.store.getState();
+        const playerID = state.ctx.actionPlayers[0];
+        const { action, metadata } = this.bot.play(state, playerID);
+
+        if (action) {
+          action.payload.metadata = metadata;
+          this.store.dispatch(action);
+        }
+
+        return action;
+      };
+    }
 
     this.reset = () => {
       this.store.dispatch(ActionCreators.reset());
@@ -100,9 +129,9 @@ class _ClientImpl {
         server,
         socketOpts,
       });
-      this.store = this.multiplayerClient.createStore(GameReducer, enhancer);
+      this.store = this.multiplayerClient.createStore(this.reducer, enhancer);
     } else {
-      this.store = createStore(GameReducer, enhancer);
+      this.store = createStore(this.reducer, enhancer);
 
       // If no playerID was provided, set it to undefined.
       if (this.playerID === null) {
@@ -138,6 +167,10 @@ class _ClientImpl {
       }
     }
 
+    if (state.ctx.gameover !== undefined) {
+      isActive = false;
+    }
+
     // Secrets are normally stripped on the server,
     // but we also strip them here so that game developers
     // can see their effects while prototyping.
@@ -147,12 +180,7 @@ class _ClientImpl {
     }
     const G = this.game.playerView(state.G, state.ctx, playerID);
 
-    if (state.ctx.gameover !== undefined) {
-      isActive = false;
-    }
-
     // Combine into return value.
-
     let ret = { ...state, isActive, G };
 
     if (this.multiplayerClient) {
@@ -173,13 +201,15 @@ class _ClientImpl {
     this.moves = createMoveDispatchers(
       this.game.moveNames,
       this.store,
-      this.playerID
+      this.playerID,
+      this.credentials
     );
 
     this.events = createEventDispatchers(
       this.game.flow.eventNames,
       this.store,
-      this.playerID
+      this.playerID,
+      this.credentials
     );
   }
 
@@ -200,6 +230,11 @@ class _ClientImpl {
       this.multiplayerClient.updateGameID(gameID);
     }
   }
+
+  updateCredentials(credentials) {
+    this.credentials = credentials;
+    this.createDispatchers();
+  }
 }
 
 /**
@@ -215,6 +250,7 @@ class _ClientImpl {
  * @param {...object} socketOpts - Options to pass to socket.io.
  * @param {...object} gameID - The gameID that you want to connect to.
  * @param {...object} playerID - The playerID associated with this client.
+ * @param {...string} credentials - The authentication credentials associated with this client.
  *
  * Returns:
  *   A JS object that provides an API to interact with the