+++ src/core/reducer.test.js
@@ -7,7 +7,7 @@
  */
 
 import Game from './game';
-import { createGameReducer } from './reducer';
+import { CreateGameReducer } from './reducer';
 import {
   makeMove,
   gameEvent,
@@ -28,16 +28,14 @@ const game = Game({
   },
 });
 
-const endTurn = () => gameEvent('endTurn');
-
 test('_stateID is incremented', () => {
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
 
   let state = undefined;
 
   state = reducer(state, makeMove('A'));
   expect(state._stateID).toBe(1);
-  state = reducer(state, endTurn());
+  state = reducer(state, gameEvent('endTurn'));
   expect(state._stateID).toBe(2);
 });
 
@@ -47,13 +45,13 @@ test('when a move returns undef => treat as illegal move', () => {
       A: () => undefined,
     },
   });
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
   let state = reducer(state, makeMove('A'));
   expect(state._stateID).toBe(0);
 });
 
 test('makeMove', () => {
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
 
   let state;
 
@@ -74,13 +72,13 @@ test('makeMove', () => {
 });
 
 test('restore', () => {
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
   const state = reducer(undefined, restore({ G: 'restored' }));
   expect(state).toEqual({ G: 'restored' });
 });
 
 test('reset', () => {
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
   let state = reducer(undefined, makeMove('A'));
   const initialState = { ...state._initial, _initial: { ...state._initial } };
 
@@ -90,13 +88,13 @@ test('reset', () => {
 });
 
 test('victory', () => {
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
 
   let state = reducer(undefined, makeMove('A'));
-  state = reducer(state, endTurn());
+  state = reducer(state, gameEvent('endTurn'));
   expect(state.ctx.gameover).toEqual(undefined);
   state = reducer(state, makeMove('B'));
-  state = reducer(state, endTurn());
+  state = reducer(state, gameEvent('endTurn'));
   expect(state.ctx.gameover).toEqual(undefined);
   state = reducer(state, makeMove('C'));
   expect(state.ctx.gameover).toEqual('0');
@@ -104,14 +102,14 @@ test('victory', () => {
 
 test('endTurn', () => {
   {
-    const reducer = createGameReducer({ game });
-    const state = reducer(undefined, endTurn());
+    const reducer = CreateGameReducer({ game });
+    const state = reducer(undefined, gameEvent('endTurn'));
     expect(state.ctx.turn).toBe(1);
   }
 
   {
-    const reducer = createGameReducer({ game, multiplayer: true });
-    const state = reducer(undefined, endTurn());
+    const reducer = CreateGameReducer({ game, multiplayer: true });
+    const state = reducer(undefined, gameEvent('endTurn'));
     expect(state.ctx.turn).toBe(0);
   }
 });
@@ -123,7 +121,7 @@ test('light client when multiplayer=true', () => {
   });
 
   {
-    const reducer = createGameReducer({ game });
+    const reducer = CreateGameReducer({ game });
     let state = reducer(undefined, { type: 'init' });
     expect(state.ctx.gameover).toBe(undefined);
     state = reducer(state, makeMove('A'));
@@ -131,7 +129,7 @@ test('light client when multiplayer=true', () => {
   }
 
   {
-    const reducer = createGameReducer({ game, multiplayer: true });
+    const reducer = CreateGameReducer({ game, multiplayer: true });
     let state = reducer(undefined, { type: 'init' });
     expect(state.ctx.gameover).toBe(undefined);
     state = reducer(state, makeMove('A'));
@@ -146,7 +144,7 @@ test('optimisticUpdate', () => {
   });
 
   {
-    const reducer = createGameReducer({ game });
+    const reducer = CreateGameReducer({ game });
     let state = reducer(undefined, { type: 'init' });
     expect(state.G).not.toMatchObject({ A: true });
     state = reducer(state, makeMove('A'));
@@ -154,7 +152,7 @@ test('optimisticUpdate', () => {
   }
 
   {
-    const reducer = createGameReducer({ game, multiplayer: true });
+    const reducer = CreateGameReducer({ game, multiplayer: true });
     let state = reducer(undefined, { type: 'init' });
     expect(state.G).not.toMatchObject({ A: true });
     state = reducer(state, makeMove('A'));
@@ -164,26 +162,26 @@ test('optimisticUpdate', () => {
 
 test('numPlayers', () => {
   const numPlayers = 4;
-  const reducer = createGameReducer({ game, numPlayers });
-  const state = reducer(undefined, endTurn());
+  const reducer = CreateGameReducer({ game, numPlayers });
+  const state = reducer(undefined, gameEvent('endTurn'));
   expect(state.ctx.numPlayers).toBe(4);
 });
 
 test('log', () => {
-  const reducer = createGameReducer({ game });
+  const reducer = CreateGameReducer({ game });
 
   let state = undefined;
 
   const actionA = makeMove('A');
   const actionB = makeMove('B');
-  const actionC = endTurn();
+  const actionC = gameEvent('endTurn');
 
   state = reducer(state, actionA);
   expect(state.log).toEqual([actionA]);
   state = reducer(state, actionB);
   expect(state.log).toEqual([actionA, actionB]);
   state = reducer(state, actionC);
-  expect(state.log).toEqual([actionA, actionB, actionC.payload]);
+  expect(state.log).toEqual([actionA, actionB, actionC]);
 });
 
 test('using Random inside setup()', () => {
@@ -202,13 +200,13 @@ test('using Random inside setup()', () => {
     setup: ctx => ({ n: ctx.random.D6() }),
   });
 
-  const reducer1 = createGameReducer({ game: game1 });
+  const reducer1 = CreateGameReducer({ game: game1 });
   const state1 = reducer1(undefined, makeMove());
 
-  const reducer2 = createGameReducer({ game: game2 });
+  const reducer2 = CreateGameReducer({ game: game2 });
   const state2 = reducer2(undefined, makeMove());
 
-  const reducer3 = createGameReducer({ game: game3 });
+  const reducer3 = CreateGameReducer({ game: game3 });
   const state3 = reducer3(undefined, makeMove());
 
   expect(state1.G.n).not.toBe(state2.G.n);
@@ -222,7 +220,7 @@ test('undo / redo', () => {
     },
   });
 
-  const reducer = createGameReducer({ game, numPlayers: 2 });
+  const reducer = CreateGameReducer({ game, numPlayers: 2 });
 
   let state = reducer(undefined, { type: 'init' });